Only rewrite ids file if not present; InitialEvidence now trying new set of commands

+++ EukMetaSanity.py
@@ -39,12 +39,15 @@ def _initialize_logging(ap: ArgParse) -> None:
 
 # Gather all files to parse that match user-passed extensions
 def _files_iter(ap: ArgParse, storage_dir: str) -> Generator[str, ArgParse, None]:
-    w = open("ids.list", "w")
+    w = None
+    if not os.path.exists("ids.list"):
+        w = open("ids.list", "w")
     for file in os.listdir(ap.args.fasta_directory):
         for ext in ap.args.extensions:
             if file.endswith(ext):
                 yield _simplify_fasta(ap, file, storage_dir, w)
-    w.close()
+    if w is not None:
+        w.close()
     return None
 
 
@@ -55,7 +58,8 @@ def _simplify_fasta(ap: ArgParse, file, storage_dir: str, w) -> str:
     record_p = SeqIO.parse(fasta_file, "fasta")
     i: int = 0
     records = []
-    w.write(file + "\n")
+    if w is not None:
+        w.write(file + "\n")
     added_records = set()
     for record in record_p:
         _record_id = "".join(choices(ascii_letters, k=16))
@@ -63,7 +67,8 @@ def _simplify_fasta(ap: ArgParse, file, storage_dir: str, w) -> str:
             _record_id = "".join(choices(ascii_letters, k=16))
         added_records.add(_record_id)
         # Write ids to stderr for user
-        w.write(_record_id + "\t" + str(record.id) + "\n")
+        if w is not None:
+            w.write(_record_id + "\t" + str(record.id) + "\n")
         # Store id
         record.id = _record_id
         records.append(record)
